Add tracker for the average commanded wheel speed
Due to the fact that the center of mass of the robot is not rigidly
fixed, the angle at which the robot can stand still without moving
in some direction changes.

This commit is the first one in the direction of adding an outer control
loop for the robot to control the speed at which it is moving via the
angle at which the robot stands.